// src/models/country.ts
export default interface country {
    code: string;
    name: string;
    symbol: string;
    iso639_1: string;
    iso639_2: string;
    nativeName: string;
    de: string;
    es: string;
    fr: string;
    ja: string;
    it: string;
    br: string;
    pt: string;
    nl: string;
    hr: string;
    fa: string;
    acronym: string;
    otherAcronyms: any[];
    otherNames: any[];
    topLevelDomain: string[];
    alpha2Code: string;
    alpha3Code: string;
    callingCodes: string[];
    capital: string;
    altSpellings: string[];
    region: string;
    subregion: string;
    population: number;
    latlng: number[];
    demonym: string;
    area: number;
    gini: number;
    timezones: string[];
    borders: string[];
    numericCode: string;
    currencies: [];
    languages: [];
    translations: [];
    flag: string;
    regionalBlocs: [];
    cioc: string;
}